split tests for test_cli_2
